requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.9
pydantic==2.10.6
pydantic-settings==2.7.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
PyMuPDF==1.25.5
rapidocr-onnxruntime==1.3.24
httpx==0.28.1
pytest==8.3.5
